Refactor UncivTextField to be a normal Widget
